pyproject.toml
[tool.poetry]
name = "cleaned-voice-agent"
version = "0.1.0"
description = ""
authors = ["Andy Spezzatti <andyspezzatti@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
redis = "^5.2.1"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
wandb = "^0.19.5"
openai = "^1.61.0"
pandas = "^2.2.3"
pydantic-settings = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
